# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Filler project's API server",
        "version": "1.0"
    },
    "paths": {
        "/api/game": {
            "post": {
                "summary": "Create new game",
                "operationId": "createGame",
                "parameters": [
                    {
                        "name": "width",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Поле создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/game/{gameId}": {
            "get": {
                "summary": "Get state of the game",
                "operationId": "getState",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game state",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "game id",
                                            "type": "string"
                                        },
                                        "field": {
                                            "properties": {
                                                "width": {
                                                    "description": "field width",
                                                    "type": "integer"
                                                },
                                                "height": {
                                                    "description": "field height",
                                                    "type": "integer"
                                                },
                                                "cells": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "color": {
                                                                "description": "cell color",
                                                                "type": "string"
                                                            },
                                                            "playerId": {
                                                                "description": "player id",
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "players": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "player id",
                                                                "type": "integer"
                                                            },
                                                            "color": {
                                                                "description": "player color",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "currentPlayerId": {
                                                    "type": "integer"
                                                },
                                                "winnerPlayerId": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неправильные параметры запроса"
                    },
                    "404": {
                        "description": "Игра с указанным ID не существует"
                    }
                }
            },
            "put": {
                "summary": "Make step in game",
                "operationId": "makeStep",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "playerId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2
                            ]
                        }
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "description": "Selected color",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "blue",
                                "green",
                                "cyan",
                                "red",
                                "magenta",
                                "yellow",
                                "white"
                            ]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ход успешно сделан"
                    },
                    "400": {
                        "description": "Неправильные параметры запроса"
                    },
                    "403": {
                        "description": "Игрок с указанным номером не может сейчас ходить"
                    },
                    "404": {
                        "description": "Игра с указанным ID не существует"
                    },
                    "409": {
                        "description": "Игрок с указанным номером не может выбрать указанный цвет"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ControllerException": {
                "title": "ControllerException",
                "description": "ControllerException model",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ControllerException"
                }
            }
        }
    },
    "security": [
        []
    ]
}